requirements.txt
langchain>=1.2.0
langchain-community>=0.4.1
langchain-core>=1.2.5
langchain-huggingface>=1.2.0
langchain-ollama>=1.0.1
langchain-qdrant>=1.1.0
langchain-text-splitters>=1.1.0
ollama>=0.6.1
pypdf>=6.5.0
pytest>=9.0.2
pytest-cov>=7.0.0
python-dotenv>=1.2.1
qdrant-client>=1.16.2
sentence-transformers>=5.2.0
streamlit>=1.52.2